requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-filter
drf-yasg
django-cors-headers
django-extensions
redis
pytest
pytest-django
coverage
# autopep8
black
gunicorn
django-environ
env
python-dotenv